Refactoring Tests Related to Spark Integration (#331)
To facilitate future expansion, we are splitting the current mixed Spark
Integration tests into the following parts:

- BaseSparkIntegrationTest abstract class: This includes the basic Spark
Catalog session creation, default setup, cleanup functions, and serves
as the base class for all Spark integration tests.
- SchemaOperationsTest: This tests schema-related operations such as
create, delete, and list.
- TableReadWriteTest: This tests table-related operations, including
read/write operations.
